Assignment 4.3/requirements.txt
networkx~=3.1
spacy~=3.5.2
docx2txt==0.8
matplotlib==3.6.2
matplotlib-inline==0.1.3
networkx==3.1
nltk==3.7
pdfrw==0.4
pikepdf==7.2.0
Pillow==9.5.0
PyPDF2==3.0.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
spacy==3.5.2
numpy==1.19.5
pandas==1.1.5
torch==2.0.1
transformers==4.29.2
scikit-learn==1.2.2
rouge~=1.0.1